requirements.txt
flask>=2.0
flask-cors
supabase
python-dotenv
